Allow homepage redirects
